+++ include/quicksr/genetic/mutation/point_mutation.hpp
@@ -31,13 +31,13 @@ namespace qsr {
  */
 class PointMutation : public BaseMutation {
 public:
-    constexpr PointMutation(float mutation_probability) :
+    constexpr PointMutation(double mutation_probability) :
         mutation_probability(mutation_probability) {}
 
     virtual std::shared_ptr<BaseMutator> get_mutator(const Config &config) override;
 
 private:
-    float mutation_probability;
+    double mutation_probability;
 };
 
 }
